Added GoalService and GoalDao. Also some minor adaptions and changes.

+++ src/main/java/de/rixtrick/demo/dao/GoalGetterDao.java
@@ -2,6 +2,6 @@ package de.rixtrick.demo.dao;
 
 import de.rixtrick.demo.model.GoalGetter;
 
-public interface GoalGetterDao extends GenericHibernateDao<GoalGetter, Long> {
+public interface GoalGetterDao extends GenericHibernateDao<GoalGetter, Integer> {
 
 }
